+++ web/klient/src/client/app/app/components/ErrorMessagePanel.spec.tsx
@@ -19,7 +19,6 @@ describe('<ErrorMessagePanel>', () => {
         errormessage: 'Error!',
       }}
       removeErrorMessage={() => undefined}
-      driftsmeldinger={[]}
     />);
 
     const div = wrapper.find(Undertekst);
@@ -36,7 +35,6 @@ describe('<ErrorMessagePanel>', () => {
         errormessage: 'Høna &amp; egget og &#34;test1&#34; og &#39;test2&#39;',
       }}
       removeErrorMessage={() => undefined}
-      driftsmeldinger={[]}
     />);
 
     const div = wrapper.find(Undertekst);
@@ -55,39 +53,11 @@ describe('<ErrorMessagePanel>', () => {
         errormessage: 'Dette er en feil',
       }}
       removeErrorMessage={() => undefined}
-      driftsmeldinger={[]}
     />);
 
     const div = wrapper.find(Undertekst);
     expect(div).to.have.length(2);
     expect(div.at(0).childAt(0).text()).to.eql('Dette er en feil ');
     expect(div.at(1).childAt(0).text()).to.eql('Feilet ');
-  });
-
-  it('skal vise Driftsmelding', () => {
-    const melding = {
-      id: '1',
-      melding: 'Dette er driftsmeldingen',
-      aktiv: true,
-      opprettet: '2020-10-10',
-    };
-
-    const wrapper = shallowWithIntl(<ErrorMessagePanel.WrappedComponent
-      intl={intl as IntlShape}
-      errorMessages={[{
-        type: EventType.REQUEST_ERROR,
-        text: 'Feilet',
-      }]}
-      queryStrings={{
-        errormessage: 'Dette er en feil',
-      }}
-      removeErrorMessage={() => undefined}
-      driftsmeldinger={[melding]}
-    />);
-
-    const div = wrapper.find(Undertekst);
-    expect(div).to.have.length(3);
-    expect(div.at(0).childAt(0).text()).to.contain('Dette er driftsmeldingen');
-    expect(div.at(1).childAt(0).text()).to.eq('Dette er en feil ');
   });
 });
