+++ web/klient/src/client/app/app/components/driftsmeldingPanel.less
@@ -1,0 +1,29 @@
+@import '@{coreModulePath}nav-frontend-core/less/_variabler';
+
+.container {
+  background-color: @navGra60;
+  color: white;
+  display: block;
+  max-height: 120px;
+  min-height: 50px;
+  overflow-x: hidden;
+  overflow-y: auto;
+  padding-bottom: 5px;
+  padding-left: 40px;
+  padding-top: 10px;
+  position: relative;
+}
+
+.wordWrap {
+  white-space: pre-line;
+}
+
+.driftsInfo {
+  height: 20px;
+  margin-right: 10px;
+  width: 20px;
+}
+
+.column {
+  display: flex;
+}
